Unbreak generated JSON files when lttng-ust spams syscalls
On an imx.6 with LTTng 2.10 I see the following events for every
lttng-ust event, when syscalls are traced too:

    syscall_entry_getcpu
    syscall_exit_getcpu
    syscall_entry_clock_gettime
    ust_event
    syscall_exit_clock_gettime

This breaks ust-event pairing, as you cannot open an event that gets
closed by the surrounding gettime. Also, the getcpu/gettime done
by lttng-ust isn't of interest to us, so we want to hide them too.

This patch detects this pattern of events and then drops the four
syscall events. With it, I can successfully generate nice graphs
from a Qt application with lttng-ust trace points on an imx.6 board.